Ruby's Float should be represented by Java's double

+++ trunk/src/com/xruby/compiler/codedom/CodePrinter.java
@@ -118,7 +118,7 @@ public class CodePrinter implements CodeVisitor {
 		result_.append("\n");
 	}
 
-	public void visitFloatExpression(float value) {
+	public void visitFloatExpression(double value) {
 		result_.append(value);
 		result_.append("\n");
 	}
